feat: (2) handle the case where an initial proposal is processed from a new space
